--- .env
@@ -1,7 +1,0 @@
-# FIREBASE
-VITE_FIREBASE_API_KEY = 'AIzaSyBemG2dk7iQ6jtmVgTU81sy5-H_HqvMML4'
-VITE_FIREBASE_AUTH_DOMAIN = 'bootcamp-project-vinted.firebaseapp.com-clone-fa336.firebaseapp.com'
-VITE_FIREBASE_PROJECT_ID = 'bootcamp-project-vinted'
-VITE_FIREBASE_STORAGE_BUCKET = 'bootcamp-project-vinted.appspot.com'
-VITE_FIREBASE_MESSAGING_ID = '239313519946'
-VITE_FIREBASE_APP_ID = '1:239313519946:web:fbcb5d48926508d982adb3'
