+++ src/components/header/navigationBar/SubCategory.tsx
@@ -4,7 +4,7 @@ import {
   MenubarSubTrigger,
 } from '@/components/ui/menubar';
 import { TypeSubCategory } from '@/types/types';
-import { buildIconPath } from '@/utils/Utils';
+import buildIconPath from '@/utils/buildIconPath';
 import SubSubCategories from './SubSubCategories';
 import { Link } from 'react-router-dom';
