+++ src/utils/constants.ts
@@ -1,6 +1,3 @@
-export enum SCREENS_BREAKPOINT {
-  Small = 640,
-  Medium = 768,
-  Large = 1024,
-}
-export const BASE_URL = `${import.meta.env.BASE_URL}`;
+const BASE_URL = `${import.meta.env.BASE_URL}`;
+
+export default BASE_URL;
